requirements.txt
# for MiniCPM4-0.5B hf inference
# torch 2.1.2 -> 2.2.0
torch==2.5.1
transformers==4.50
accelerate==0.26.0
gradio==5.11.0
safetensors<0.5.0

# for evaluation
rouge-score==0.1.2
rouge_chinese==1.0.3

# for tokenization
tokenizers

# for openai api inference
# openai==1.17.1
# tiktoken==0.6.0
# loguru==0.7.2
# sentence_transformers==2.6.1
# sse_starlette==2.1.0

# for MiniCPM-V hf inference
Pillow==11.3.0
timm==0.9.16
sentencepiece==0.2.0

# for finetune
peft==0.6.2
ruamel_yaml==0.18.5
jupyter==1.0.0
datasets==2.16.1
deepspeed==0.14.0

# Fix version conflicts
numpy<2.3.0,>=2
websockets>=13.0,<15.1.0
# torchaudio 2.1.2 -> 2.2.0
torchaudio==2.5.1
# git+https://github.com/michaelchen-lab/caft-llm.git 